created JsonManager File with FileManager interface
