+++ .gitignore
@@ -20,3 +20,5 @@ tests/__pycache__/diagnosticAid.cpython-39.pyc
 tests/__pycache__/mockData.cpython-39.pyc
 tests/__pycache__/diagnosticAid.cpython-39.pyc
 tests/__pycache__/mockData.cpython-39.pyc
+main/test_gathering_json.py
+main/__pycache__/fileManager.cpython-310.pyc
