+++ main/mainProgram.py
@@ -9,8 +9,30 @@ import os,sys
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.dirname(SCRIPT_DIR))
 from main.goldPricedInSilver import GoldPricedInSilver
-from main.setup import setUpJsonFiles
 from os.path import exists as file_exists
+from main.jsonManager import JsonManager
+
+def setUpJsonFiles():
+    mainFileName = 'GoldPricedInSilver.json' 
+    cachedFileName = 'cachedData.json'
+
+    # Header for GoldPricedInSilver.json
+    mainHeader = {}
+    mainHeader['GoldPricedInSilver'] = []
+
+    # Headers for cachedData.json
+    cachedDataHeaders = {}
+    cachedDataHeaders['sum'] = 0
+    cachedDataHeaders['priceList'] = []
+    cachedDataHeaders['priceListNegative'] = []
+
+    """
+    Using JsonManager to create 'GoldPricedInSilver.json' and
+    'cachedData.json.
+    """
+    mainJsonFile = JsonManager(mainFileName, **mainHeader)
+    cachedJsonFile = JsonManager(cachedFileName, **cachedDataHeaders)
+
 
 if __name__ == '__main__':
     headerAndfileNames = ['GoldPricedInSilver.json', 'GoldPricedInSilver', 'cachedData.json']
@@ -20,3 +42,16 @@ if __name__ == '__main__':
 
     goldSilverPricing = GoldPricedInSilver(*headerAndfileNames)
     goldSilverPricing.addNewPrice()
+
+
+# TODO: What's the difference between addNewPrice() and saveNewData()
+# We may need to delete updateData.py based on what we can figure out.... 
+"""
+    headerAndfileNames = ['GoldPricedInSilver.json', 'GoldPricedInSilver', 'cachedData.json']
+
+    if file_exists('GoldPricedInSilver.json') and file_exists('cachedData.json'):
+        goldSilverPricing = GoldPricedInSilver(*headerAndfileNames)
+        goldSilverPricing.saveNewData()
+
+
+"""
